Respond to comments on ingest
